requirements.txt
AWSIoTPythonSDK>=1.4.6
boto3>=1.9.126
botocore>=1.12.126
configparser>=3.7.4
Flask>=1.1.1
Flask-Bootstrap>=3.3.7.1
flask-nav>=0.6
MarkupSafe>=1.1.1
paho-mqtt>=1.4.0
pyserial>=3.4
python-dateutil>=2.8.0
s3transfer>=0.2.0
urllib3~=1.25
Werkzeug>=0.16.0
wifi>=0.3.8
asn1crypto>=1.3.0
Click>=7.0
cryptography>=2.1.4
dominate>=2.3.5
enum-compat>=0.0.2
funcsigs>=1.0.2
idna>=2.6
ipaddress>=1.0.17
itsdangerous>=1.1.0
keyring>=21.1.0
keyrings.alt>=3.4
mock>=2.0.0
pbkdf2>=1.3
pbr>=3.1.1
pycryptodome>=3.9.8
setuptools>=45.1.0
svgwrite>=1.4
pygame~=1.9.6
pip>=20.0.2
requests>=2.24.0
flask-requests>=0.0.14
flask-request-params>=0.3.0
retry>=0.9.2
tabulate>=0.8.7
numpy
matplotlib==3.4.3
wiringpi>=2.60.0; sys_platform == 'linux'
watchtower>=0.8.0
robotframework>=3.2.2
psutil>=5.7.3
PyYAML>=5.4
Pillow==8.3.2
